application restuctured correctly
