+++ my-app/app/(auth)/register/page.tsx
@@ -1,6 +1,6 @@
 import { Metadata } from 'next';
 import React from 'react';
-import Register from '@/app/components/auth/Register';
+import Register from '@/components/auth/Register';
 
 export const metadata: Metadata = {
     title: "Register",
