+++ my-app/app/(auth)/resetpassword/page.tsx
@@ -1,0 +1,13 @@
+import { Metadata } from 'next';
+import React from 'react';
+
+export const metadata: Metadata = {
+    title: "Reset Password",
+    description: "reset password page"
+};
+
+export default function ResetPasswordPage() {
+    return (
+        <div>ResetPasswordPage</div>
+    )
+}
