+++ my-app/app/api/register/route.ts
@@ -1,0 +1,4 @@
+export async function POST(req: Request): Promise<Response> {
+
+    return Response.json("Register done !");
+}
