+++ my-app/app/dashboard/dashboardnative/@tabletwo/default.tsx
@@ -1,11 +1,11 @@
-import type { ProductsProps } from '@/app/lib/definitions';
+import type { ProductsProps } from '@/lib/definitions';
 import React, { Suspense } from 'react';
 import Link from 'next/link';
-import TablePage from '@/app/components/TablePage';
-import BestSellersChart from '@/app/components/graphs/BestSellersChart';
-import StockChart from '@/app/components/graphs/StockChart';
-//import { products } from '@/app/lib/products';
-import Loader from '@/app/components/Loader';
+import TablePage from '@/components/TablePage';
+import BestSellersChart from '@/components/graphs/BestSellersChart';
+import StockChart from '@/components/graphs/StockChart';
+//import { products } from '@/lib/products';
+import Loader from '@/components/Loader';
 
 export default async function TableTwoDefault() {
