+++ my-app/app/dashboard/dashboardnative/default.tsx
@@ -1,7 +1,7 @@
 import React, { Suspense } from 'react';
-import Header from '@/app/components/Header';
-import Menu from '@/app/components/Menu';
-import Loader from '@/app/components/Loader';
+import Header from '@/components/Header';
+import Menu from '@/components/Menu';
+import Loader from '@/components/Loader';
 
 export default function DashboardDefault() {
     return (
