+++ my-app/app/email/page.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function EmailPage() {
+    return (
+        <div>EmailPage</div>
+    )
+}
