+++ my-app/app/products/layout.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import HeaderProducts from '@/app/components/products-and-cart/HeaderProducts';
+import HeaderProducts from '@/components/products-and-cart/HeaderProducts';
 
 export default function ProductLayout({children}: {children: React.ReactNode}) {
     return (
