+++ my-app/app/products/page.tsx
@@ -1,8 +1,8 @@
 import type { Metadata } from 'next';
 import React from 'react';
 import { dehydrate, HydrationBoundary, QueryClient } from "@tanstack/react-query";
-import { getProductsData } from "@/app/utils/api-request";
-import AllProducts from '@/app/components/AllProducts';
+import { getProductsData } from "@/utils/api-request";
+import AllProducts from '@/components/AllProducts';
 
 export const metadata: Metadata = {
   title: {
