+++ my-app/components/TableGraphs.tsx
