+++ my-app/components/auth/PasswordForgot.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export default function PasswordForgot() {
+    return (
+        <div>PasswordForgot</div>
+    )
+}
