+++ my-app/components/graphs/BarChartBrowser.tsx
