+++ my-app/components/graphs/LineChart.tsx
