+++ my-app/components/products-and-cart/Card.tsx
@@ -1,9 +1,9 @@
 "use client";
 
-import { ProductsProps } from '@/app/lib/definitions';
+import { ProductsProps } from '@/lib/definitions';
 import React from 'react';
-import { useStore } from '@/app/lib/store';
-import usePersistStore from '@/app/helpers/usePersistStore';
+import { useStore } from '@/lib/store';
+import usePersistStore from '@/helpers/usePersistStore';
 import Link from 'next/link';
 import Image from 'next/image';
 import Loader from '../Loader';
