+++ my-app/components/products-and-cart/HeaderProducts.tsx
