+++ my-app/lib/products.ts
