+++ my-app/lib/store.ts
