+++ my-app/utils/api-request.ts
@@ -1,4 +1,4 @@
-import type { GeoLocationData, ProductsProps } from "@/app/lib/definitions";
+import type { GeoLocationData, ProductsProps } from "@/lib/definitions";
 import { revalidatePath } from "next/cache";
 
 export const ApiPublicIp = async (): Promise<{data: {ip: string;}}> => {
