+++ my-app/utils/generate-version.ts
